[Update] CurrencyNotFoundException class moved to exceptions package
