+++ src/main/java/org/currencygoldexchangeapp/exceptions/CurrencyNotFoundException.java
@@ -1,4 +1,4 @@
-package org.currencygoldexchangeapp.services;
+package org.currencygoldexchangeapp.exceptions;
 
 public class CurrencyNotFoundException extends RuntimeException {
     public CurrencyNotFoundException(String message) {
